pyproject.toml
[project]
name = "tg-whatsapp-router"
version = "0.1.0"
description = "Monitor Telegram public channels and forward messages to WhatsApp"
authors = [{name = "Alexander Domoryonok"}]
license = {text = "MIT"}
keywords = ["telegram", "whatsapp", "message-router", "telethon"]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptg>=0.5.2",
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "telethon>=1.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
